pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "fpdf>=1.7.2",
    "google-generativeai>=0.8.4",
    "pdfplumber>=0.11.6",
    "pillow>=11.1.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "streamlit>=1.44.1",
    "textstat>=0.7.5",
]
